requirements.txt
pymilvus>=2.5.0

sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0

langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
langchain-google-genai>=1.0.0

numpy>=1.21.0
pandas>=1.5.0

openai>=1.0.0

tqdm>=4.64.0

# Add PyYAML for configuration management
PyYAML>=6.0

# DSPy for automatic prompt optimization
dspy-ai>=2.4.0 